+++ src/noddos.cxx
@@ -90,7 +90,7 @@ int main(int argc, char** argv) {
 		openlog(argv[0], LOG_NOWAIT | LOG_PID | LOG_PERROR, LOG_UUCP);
 	}
 	Config config(configfile, debug);
-	InterfaceMap ifMap(config.LanInterfaces,config.WanInterfaces, true);
+	InterfaceMap ifMap(config.LanInterfaces,config.WanInterfaces, config.Debug);
 
 	if (daemon) {
 		daemonize(config);